Crate transcript

Module for PolyIOP transcript. TODO(ZZ): move this module to HyperPlonk where the transcript will also be useful. TODO(ZZ): decide which APIs need to be public.

Macros

to_bytes

Takes as input a struct, and converts them to a series of bytes. All traits that implement CanonicalSerialize can be automatically converted to bytes in this manner.

Structs

IOPTranscript

An IOP transcript consists of a Merlin transcript and a flag is_empty to indicate that if the transcript is empty.

Enums

TranscriptError

A enum specifying the possible failure modes of the Transcript.